+++ includes/libraries/guzzlehttp/Ring/Exception/CancelledException.php
@@ -1,7 +1,0 @@
-<?php
-namespace MAGuzzleHttp\Ring\Exception;
-
-/**
- * Marker interface for cancelled exceptions.
- */
-interface CancelledException {}