+++ includes/libraries/guzzlehttp/Ring/Exception/CancelledFutureAccessException.php
@@ -1,4 +1,0 @@
-<?php
-namespace MAGuzzleHttp\Ring\Exception;
-
-class CancelledFutureAccessException extends RingException implements CancelledException {}