+++ includes/libraries/guzzlehttp/Ring/Exception/ConnectException.php
@@ -1,7 +1,0 @@
-<?php
-namespace MAGuzzleHttp\Ring\Exception;
-
-/**
- * Occurs when the connection failed.
- */
-class ConnectException extends RingException {}