+++ includes/libraries/guzzlehttp/Ring/Exception/RingException.php
@@ -1,4 +1,0 @@
-<?php
-namespace MAGuzzleHttp\Ring\Exception;
-
-class RingException extends \RuntimeException {};