+++ includes/libraries/guzzlehttp/Stream/Exception/CannotAttachException.php
@@ -1,4 +1,0 @@
-<?php
-namespace MAGuzzleHttp\Stream\Exception;
-
-class CannotAttachException extends \RuntimeException {}